requirements.txt
appdirs==1.4.3
enum34==1.1.6
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
vk-api==8.6
